+++ src/pages/_app.tsx
@@ -11,7 +11,7 @@ import '@/shared/styles/index.scss'
 import '../shared/ui/toast/toast.css'
 import 'nprogress/nprogress.css'
 import { StoreProvider } from '@/providers/store-provider/store-provider'
-import { useLoader } from '@/shared/hooks/useLoader'
+import { useLoader } from '@/shared/hooks/use-loader'
 import { ToastNotify } from '@/shared/ui/toast/toast'
 
 export type NextPageWithLayout<P = {}, IP = P> = NextPage<P, IP> & {
